+++ NEWS.md
@@ -4,6 +4,8 @@ This is a major release, adding methods for linear/quadratic discriminant analys
 
 * Added `predict_discrim()` as a wrapper to `MASS::predict.lda()` returning a data.frame
 * Added `plot_discrim()` to plot decision regions for discriminant analysis using contours or background tiles
+* Added `cor_lda()` to calculate structure correlations between variables and discriminant scores
+* Added `scores.lda()` method to extract discriminant scores
 * Added `reflect()` generic to allow axis reversal for candisc and cancor objects
 * Improved documentation of a number of functions
 * Added additional plots to dataset documentation